feat(COR-1954) Split up and combine into module
