+++ packages/app/src/components/metadata/components/items/metadata-icon.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import css from '@styled-system/css';
+
+export const MetadataIcon = styled.span(() =>
+  css({
+    minWidth: '1.8rem',
+
+    svg: {
+      height: '15px',
+      width: 'auto',
+    },
+  })
+);
